requirements.txt
# Backend Server Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Printer Management
psutil==5.9.6
pywin32==306

# Network Discovery
zeroconf==0.131.0
requests==2.31.0

# File Handling
Pillow==10.1.0
PyPDF2==3.0.1

# Database (Optional for job history)
# sqlite3 is included in Python standard library

# Web Interface
jinja2==3.1.2
# static files served by FastAPI built-in

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pyyaml==6.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
pyinstaller==6.2.0